feat: 移动端主题设置入口移入侧边栏并弹出选项
- 新增 AccentThemePanel 组件：主题色色板 + 深/浅色切换，桌面弹层与移动端弹窗共用
- 移动端（≤600px）：侧边栏底部原深色模式按钮改为主题设置入口，点击弹出居中弹窗（遮罩点击关闭）；顶栏调色板按钮及锚定弹层隐藏
- 桌面端行为不变：顶栏调色板按钮 + 右侧锚定弹层
- 修复点击侧边栏按钮即开即关的问题（点击外部关闭逻辑将触发按钮视为内部）
- 修复移动端同时弹出两个弹窗的问题（移动端隐藏桌面端弹层容器）
